LCM: Revising regression tests for single and multislip hardening
The single slip case was not quite right in terms of the acceptable residual. Updated.
In addtion, we also updated the gold file for the multislip case.

These changes reflect a new predictor for the slip at t_{n+1}. More work to do here regarding
a minimum number of iterations for solution.