Fixing NSVortexShedding_Blocked example
This problem tests the Blocked numbering scheme as compared
to the default Interleaved ordering. The two tests also differed
in by using Aztec instead of Belos in the eigensolver.
I verified that the failure could be solved by tightening the
tolerances. In the end I switched the Block version of the
problem to also use Belos, so that there is only one difference
in the Blocked test.