LCM: Changes to the CP algorithm required updating gold files.
For some tests, setting a minimum number of constitutive iterations improved the
residual and caused small changes in the initial stages of slip. Rather than adjust
the floor, I am just blessing new tests. For both cases, the residual is reduced.

In addition - I'm now copying the python scripts for checking quantities into the test directory.

Finally - changed the minimum number of iterations to 2. This is a bit cheezy but it bought me
another order of magnitude in testing. We will do something a bit better in the future.