LCM: Committing new test for a pressure-dependent concentraiton boundary condition
For the applied concentration boundary condition to respect the chemical potential of the solid
the concentration must be a function of the pressure. Because we often perform an L2 projection
of the Kirchhoff pressure, we have the pressure at the nodes.

Jake implemented the boundary condition. This test verifies it  with a gold file and the
analytical expression built into a python script, plot_hard_implicit.py, for checking.

I will do a little more clean up and copy the python script into the test directory.

Thanks again to Jake for getting this rolling.