LCM: Getting ready case which shows distinct differences between Albany & Albanyt with ILU
Refreshing example to show differences between Albany and AlbanyT

Albany with ILU has no issues. AlbanyT bangs. Belos struggles. Does not achieve equivalent residual.

This problem is multiphysics and very challenging. Units are scaled to mitigate condition number issues.
Surface elements present. There is a massive difference in the mesh size. All that said, Albany is
solving the system and the residual is bottoming out around 1e-4. AlbanyT does no better than 1e-2 and
the process is less than comforting.

This residual tolerance is a little loose in the beginning of the simulation. This is "fixed" with a
minimum nunmber of iterations (4). In Albany, the residual bottoms out. Not the case in AlbanyT...

Now including a CMakeLists.txt file so that all the files will be copied into the build directory for testing.

Albany: inputAllLocalHarder.xml

              <Parameter name="Preconditioner Type" type="string" value="Ifpack"/>
              <ParameterList name="Preconditioner Types">
                <ParameterList name="Ifpack">
                  <Parameter name="Overlap" type="int" value="2"/>
                  <Parameter name="Prec Type" type="string" value="ILU"/>
                  <ParameterList name="Ifpack Settings">
                    <Parameter name="fact: drop tolerance" type="double" value="0"/>
                    <Parameter name="fact: ilut level-of-fill" type="double" value="1"/>
                    <Parameter name="fact: level-of-fill" type="int" value="1"/>
                  </ParameterList>
                </ParameterList>
              </ParameterList>

AlbanyT: inputAllLocalHarderT.xml

              <Parameter name="Preconditioner Type" type="string" value="Ifpack2"/>
              <ParameterList name="Preconditioner Types">
                <ParameterList name="Ifpack2">
                  <Parameter name="Overlap" type="int" value="2"/>
                  <Parameter name="Prec Type" type="string" value="RILUK"/>
                  <ParameterList name="Ifpack2 Settings">
                    <Parameter name="fact: drop tolerance" type="double" value="0"/>
                    <Parameter name="fact: ilut level-of-fill" type="double" value="1"/>
                    <Parameter name="fact: level-of-fill" type="int" value="1"/>
                  </ParameterList>
                </ParameterList>
              </ParameterList>