LCM: Kokkoization J2MiniSolver.
First cut at skeleton. Defining tags, execution spaces and so on
is done. But in defining the functor that will calculate the J2
residual there are some design decisions to make.

Other Albany evaluators that are Kokkoized store all their data
in member variables. Not sure that for constitutive models in LCM
we want to/can do that. Discuss with Jake and Irina D. how to proceed.