Finish adding compile option ENABLE_STOKHOS
Configuration option -DENABLE_STOKHOS:BOOL=OFF allows
compiling without the Trilinos Stokhos package. The default is
ENABLE_STOKHOS:BOOL=ON.

This is a step towards compiling Albany Epetra free. In the near term,
we are debugging a GO=long issue and it is helpful to reduce the amount
of 32 bit code (both Tpetra and Epetra) to assist in finding a couple of
outstanding issues.