MOR: modify tests to run after SolutionDot changes
The MOR tests read in an exodus file containing the POD basis.  With the
changes in Albany related to SolutionDot, Albany expects the basis file to
contain data for SolutionDot.  Luckily the basis file had an extra field,
labeled as "solution_height" that was all zero.  To get these tests to run,
the xml files have been modified to treat "solution_height" as SolutionDot.

This is just a hack to get these tests to run again.  The POD basis files
should be regenerated, but this may require changes to AlbanyRBGen.