Fix in STKDiscretization
The way we built the projectors (from the discretization to the side
discretizations) was buggy. In particular, it worked fine in every case
except for the basal mesh of ExtrudedSTKMeshStruct with column-wise
ordering. In fact, in that case, the GID of the nodes in the basal mesh
is NOT the same as the GID of those nodes in the extruded mesh.

To fix it, we used the sideToSideSetCell and the sideNodeNumerationMap
in the discretization, to correctly match each node in the mesh with
the corresponding one in the side mesh.