Small change in DiscretizationFactory
I added a method called 'createMeshStruct'. This method is now called by
'createMeshSpecs' (rather than have all the if/else inside it).
The advantage is that createMeshStruct does not alter the variables inside
the factory, and can therefore be called multiple times. This is exploited
when Extruded meshes have to be created, since there are two mesh structs
to create: basal mesh and extruded mesh. Now we can use 'createMeshStruct'
to create both of them. This also allowed to remove the creation of the
basal mesh inside ExtrudedSTKMeshStruct. Instead, the constructor of
ExtrudedSTKMeshStruct now requires an additional input parameter, which
is the basal mesh struct.