Modifications related to SG
Some members types of a few classes are related to Stokhos types. The declaration (and use)
of these types should happen only when ALBANY_STOKHOS is defined.
In addition, the tests MPNIQuad2D and SGQuad2D always require Stokhos (even for the
AlbanyT/Albany tests), since in the input file we require KL expansions (whose code is
compiled only when ALBANY_STOKHOS is defined).