LCM: Added journal file for convergence study of Schwarz.
This is a single domain mesh with ~200,000 hex elements.
The elements have more or less an aspect ratio of one after some
manual iteration. This mesh is intended to be used to compute
a single mesh reference solution to determine the error introduced
by the Schwarz iteration and separate it from the discretization error.