LCM: adding capability to scale the rows of the Jacobian/residual that correspond to Dirichlet dofs instead of vice versa.  This option is off by default.  To turn it on, it is necessary to add the "Scale BC Dofs" line under the "Scaling" parameter list in the input file and set this option to true, e.g.,
 <ParameterList name="Scaling">
    <Parameter name="Scale" type="double" value="1.0e4"/>
    <Parameter name="Scale BC Dofs" type="bool" value="true"/>
  </ParameterList>

Currently, this option is only available for Mechanics problems.  It can
be extended to any Albany problem simply by adding a line having to do
with offsets to the problem's *.hpp and *.cpp files.

I have tested the new options on CrystalPlasticity/MultiSlipHard and
CrystalPlasticity/RubiksCube and it works in serial and parallel.
More testing would be good.

The new option does not seem to work correctly on Schwarz problems
when run in parallel.