LCM: Fixing equivalent plastic strain and modifying Rubik's cube for Hojun
The eqivalent plastic strain is generally

\dot{\epsilon_{p} = sqrt(2/3 * D_{p} :: D_{p}})

where D_{p} is the symmetric part of the plastic velocity gradient L_{p}.

We had in the code:

\dot{\epsilon_{p} = 2/3 * sqrt(D_{p} :: D_{p}})

I noticed this when putting together a test for eqps.

eps \sim ln(l/l_{0} - cauchy_stress_component/effective_young's modulus

For a cubic system, you can approximate Young's modulus by just fitting the linear portion of the elastic loading.

Give those assumptions, both metrics for eqps are spot on (in the context of plotting).

While helpful, this is not verficiation to numerical precision. We will get that done soon. I need to help Hojun.

Equivalent plastic strain is only helpful for visualization. It is a derived quantity.

Tests need to be modified. That's next.