Aeras: finishing hooking in of new Aeras::ComputeAndScatterJac class.
The Jacobian specialization of evaluateFields in Aeras::ComputeAndScatterJac
needs to be rewritten to remove AD.  This function already has wBF, BF, wGradBF,
and GradBF, all the fields needed to compute the local mass and laplacian.
Right now, the guts of this function are copied from Aeras_ScatterResidual_Def.hpp,
so it gets called, and the code runs w/ no problems.  The same should be the case
w/ the reimplementation of evaluateFields.  Once the reimplementation is done,
we can remove all the un-needed dependencies b/w the ---- in the
constructEvaluators<PHAL::AlbanyTraits::Jacobian> function in Aeras_HydrostaticProblem.cpp.