Bug fix in "Scale BC Dofs" option for system scaling.  Before, scaling was not done properly for Dirichlet BCs that set all the dofs at once (including Schwarz); this should be fixed now.
Although I have verified that the scaled Schwarz submatrices are correct
w/ this bug fix, convergence for Schwarz with "Scale BC Dofs" set to true is not
the same as when this is set to false.  Need to talk to Alejandro if this might
have something to do with the Schwarz formulation (although I do not see that it would).