Added a user-defined NOX status test that can be flagged from within a model evaluator.
The status test allows constitutive models to set the NOX status to Failed, which
in turn triggers a global load step reduction via Piro for continuation problems
with adaptive time stepping turned on.

Currently the user defined status test is passed only to the CrystalPlasticity
material model.  This model has a state update routine that can occasionally fail
for various reasons, in which case it's desirable to reduce the load step.