AAAACK! Putting exit(-1) for all problems that Use Serial Mesh
A change in Trilinos is causing all problems that have the
Use Serial Mesh option to hang. This was preventing our testing.

I have put an "exit(-1);" in the code when this is accessed
so at least our testing progresses with  failed  instead
of stalling until test time out.

Need to fix this for real, ASAP.