Re-added "Use Serial Mesh" capability to partition mesh in the code.
I implemented Greg Sjaardema's option in STKIO and Ioss to
partition a single exodus mesh for parallel runs. Previously,
we had been using stk_rebalance for this capability, but
some change in Seacas broke this a few days ago.

For now I hardwired the partitioning to RIB (recursive
intertial bisection) but it could be made an option.

Tests should pass again.