Removed the requirement that Mechanics problems have Piro->NOX-Solver Options and Prio-NOX->Status Tests ParameterLists.
This requirement was recently added so that a user-defined NOX status test and a
NOX pre/post operator could be added by the Mechanics Problem.  With these changes,
the status test and pre/post operator will be added only if the associated
ParameterLists are present.  If they are not present, the status test and pre/post
operator are simply not added and the simulation proceeds without them.

 LocalWords:  src Untracked djl txt cmake albany intel linux dbg