Some changes for layered states.
A layered state is useful for extruded meshes, to save the value of a 3D
state at each mesh layer as a vector field in the basal mesh. The issue
is that the number of layers is not linked to either the dimension of the
mesh nor the dimension of vector quantities (VecDim). Therefore, a new tag
had to be introduced. The corresponding layout is vector-dependent, and the
user is responsible to create a proper DataLayout for each layered field.

Furthermore, in order to store the coordinates of each layer, we needed to
save a global variable in the mesh. This required to change the way the
exodus output is processed. In particular, 'process_output_request' does
not process global variables, and had to be replaced with 'write_defined_output_fields'
and 'write_global' (if any).