PHAL: added a response type (with its evaluators) and an evaluator.
Added a response that computes the L2 norm of a mismatch on a sideset. The
target field can be either a fixed field (ParamScalarT, with no derivative)
or a solution field (ScalarT, with Sacado derivatives). The first is useful
to match data, the second can be handy to compute misfit between two solutions
on an interface.

Added an evaluator that interpolates the gradient of a vector field on a side.