PHAL: Diversification of side interpolation evaluators for different ScalarT types.
We have two version of each interpolation evaluators, depending on the type
of the underlying scalar: ScalarT or ParamScalarT. The first uses Sacado
derivatives for Jacobian, Tangent, etc., while the second keeps double as
underlying scalar, since the field is simply a parameter (unless EvalT
has ParamScalarT that is of Fad type, such as when mesh/parameters depend
on the solution).

This distinction allows to not compute all the derivatives for input fields
and solution-independent fields during EvalT=Jacobian,Tangent,etc., saving
both time and storage.