FELIX: Introduced a static version of SharedParameter
This is very similar to PHAL::SharedParameter (and eventually may/should
replace it), but with a difference: the stored value is static. This ensures
that the stored value is the same in all the instantiation of the class inside
different field managers. I experienced some problem with ROL, where it seemed
like the stored value was not updated in every instantiation of the class
in every field manager (in particular the response field manager).
With this implementation, instead, the value is the same across all field
manager.

However, this required some additional templated parameters, since we don't
want the SharedParameter objects relative to different parameters to store
the same value. To avoid this, I introduced two templated arguments: an
EnumType and a value of that EnumType. I thought about simply using one
extra templated parameter, an int, but that was less wordy, while an enum
helps. The drawback is that one has to define the enum for his/her problem
before creating any SharedParameter object.

Example:

enum MyParams { Alpha, Beta, Gamma};

SharedParameter<EvalT,Traits,MyParam,Alpha> evalAlpha;
SharedParameter<EvalT,Traits,MyParam,Alpha> evalAlphaAgain;

These two objects will always store the same value for alpha, and one does
not need to modify the value in both.