Refactoring the way we handle different scalar type for evaluators.
For non-Residual evaluation types. The interpolation evaluators were
modified to work for both solution-dependent fields and "parameter" fields
(that do not need sacado derivatives). To do this, we have added a third
parameter (the ScalarType) to the evaluator. Since the macro in AlbanyTraits
to instantiate templates did not work with three template arguments, these
evaluators were divided in a hyerarchy: XXXBase (with 3 template arguments),
XXX (the "standard" one, using EvalT::ScalarT as scalar type) and XXXParam
(which used EvalT::ParamScalarT as scalar type). However, there are few
drawbacks:

1) we introduce inheritance, when all XXX and XXXParam do is to force
   ScalarType to be of a desired type
2) EvaluatorUtils needed 2 different methods for each evaluators, namely,
   createXXX and createXXXParam.

To overcome this, we did the following:

1) Introduced a new PHAL macro, to instantiate a templated class with 3
   template arguments, using all the scalars in EvalT as third argument
2) EvaluatorUtils is now EvaluatorUtilsBase and takes 3 arguments, with
   the 3rd one being the scalar type (just as the evaluators). We added
   a template alias EvaluatorUtils, which is an alias of the "usual"
   EvaluatorUtils, with scalar type equal to EvalT::ScalarT (which allows
   to not change old code that uses EvaluatorUtils<EvalT,Traits>).
   In addition, there are 2 helper methods, getPSUtils and getMSUtils,
   which return, respectively, an EvaluatorUtilsBase with scalar type
   EvalT::ParamScalarT and EvalT::MeshScalarT.

Now, if you need to create the interpolation XXX with scalar type
EvalT::ScalarT, you use the same syntax: evalUtils.createXXX(...).
To create the evaluator with scalar type EvalT::ParamScalarT, you do
evalUtils.getPSUtils().createXXX(...).