Added PHAL evaluator
Added an evaluators that adds noise to a field. The noise can be either
absolute or relatie to the input field, and can have uniform or normal
distribution. It is also possible to use this evaluator without an input
field, to simply generate random vectors. Furthermore, one can decide
whether the added noise is the same at every evaluation or not.