Mod in FELIX SharedParameter
The method to set the nominal value for the parameter now takes 2 arguments: a parameter list
and a double. Before resorting to the input double, the method scans the input list. This is
supposed to be the 'Parameter' sublist of the problem section in the input file. Therefore,
the method looks for the parameter vector with the right name; if found and if there are
nominal values for that parameter it loads the values, otherwise uses the input default value.