Major modification with how side sets layouts are handled
Before, a struct Layouts contained both volume and surface layouts. If more
than one sideset was present, one had to create two Layouts structure, which
would have the same volume part, but potentially different surface layouts.
Moreover, the Layout struct would contain twice as many structures, having
a side version for (almost) any volume layout.

Now things are different. A Layout struct stores a map to RCP pointers to
side set Layouts. A Layout strucutre has the same attributes, regardless of
whether it is a volume or surface one. Clearly, the layouts will be different
in case it is relative to a side set (namely, will contain the Side tag).

Among the advantages, one only needs to pass one Layout structure to an
evaluator, even if that evaluator handles two or more sidesets. Inside the
evaluator, one simply access the side set Layouts stored inside the Layouts
struct (perhaps after checking that they are defined).

Warning: when one sees something like 'dl->node_scalar', it is not clear
if the layout is <Cell,Node> or <Cell,Side,Node>. The user must be careful
when handling such layouts, namenly he has to know if it is a side set layout
or not. He can easily check this inspecting the variable 'dl->isSideLayouts',
which is true if the Layouts are defined on a side set (and hence contain
the Side tag) and false otherwise.