Aeras: adding configure option
-D ENABLE_AERAS_IMPLICIT_HS:BOOL=ON \

which if off by default.  This should be turned on if one wishes to
run implicit 3D hydrostatic problems in Aeras, which will not work by
default due to code memory optimizations recently made.  Note that
when this option is on, the explicit problems will not be as memory efficient,
as AD will be used to compute the Jacobian matrix.