Aeras: changing "Aeras Hydrostatic Baroclinic Instabilities Perturbed" initial condition to "Aeras Hydrostatic Baroclinic Instabilities", and making it so this initial condition can encompass both the perturbed and unperturbed version of the problem.  The 3rd entry of the "Function Data" is now a boolean which tells the code what initial condition to set.  A value of 1 (= true) means perturbed; a value of 0 (=false) means unperturbed.
The older "Aeras Hydrostatic Baroclinic Instabilities Unperturbed" IC still
exists but should be removed in my opinion.  It is redundant with the "Aeras Hydrostatic Baroclinic Instabilities"
IC and I think has some erroneous values when compared to the Jablonowski paper.
Need to talk to Tom about this IC and whether it can be removed.