Aeras: rewrote Kokkos functor implementation of Aeras::ComputeBasisFunctions class.  The older version had been commented out in the code for almost a year and did not work correctly.  In the new implementation, I replaced routines calculating basis functions, Jacobian determinant, etc. with Interepid2 calls, since thewew routines are now Kokkos-ized in Intrepid2.
The HOMMEMAP implementation of the jacobian currently does not exist in the Kokkos
functor version of this class, nor does the calculation of GradGradBF and wGradGradBF.
Need to check with Oksana if the HOMMEMAP implementation needs to be Kokkos-ized,
and also about whether GradGradBF and wGradGradBF are needed.  I think they are a part
of an old, incorrect version of hyperviscosity and should be removed from the code
entirely to prevent confusion.