Mods to PHAL_DOF* evaluators
We conformed the cell DOF* evaluators to the side ones, in that we added
a template argument that establish the scalar type of the interpolated
field. The mod can help to decrease storage/compuatational costs for
input fields that are constant with respect to the solution/parameters,
and therefore do not require derivatives (no Fad structures to fill).