Added --combine_blocks option to imprint_orientation_on_mesh.py
The intended use case is 1) create a polycrystal genesis mesh in which
each grain is a block, 2) create a text file containing a rotation
matrix for each block, 3) run this script with the --combine_blocks option.

The script records the orientations as element attributes, and then combines
the blocks into a single block.  So the per-block orientations are preserved,
but the actual blocks are combined into one block.  Having a single block
circumvents usability issues and may improve performance in Albany/LCM.