pumi : move mesh struct variable init to constructor
The mesh struct initialization process goes like this:
constructor for derived class (PUMI,SIM) -> APFMeshStruct->init().
Initializing the variable for restartWriteStep in the call to init()
overwrote the value set in the derived class and broke a regression
test. This commit should fix that by leaving it up to the derived
classes to initialize this value.