PUMI: safety check for mismatched layout and neq
The SCOREC test that is currently segfaulting on
CDash can be traced to a mismatch between the
number of equations specified by the layout in
the input.xml file and the value of the "neq" variable.

This commit catches such a mismatch as a Teuchos
exception to prevent invalid memory accesses later on.
Both adaptive and non-adaptive Dynamics tests trigger
this assertion.

Other formatting fixes also included.