SCOREC : update dynamics examples
This commit addresses the mismatch caught by bd5e3e9
for two dynamics tests.

Pressure is only set as a degree of freedom for mechanics
problems if the parameter list "Stabilized Pressure"
exists with a "Variable Type" set to DOF. i.e The old
inputs for these problems were specifying an extra
degree of freedom in the solution vector layouts.