LCM: Change interface for parallel constitutive models in order to simplify constitutive model development.
Added a new base class for parallel kernels. This provides helper functions for interfacing with the current constitutive model.

Fixed a bug in which addStateVar ignored the "old" flag.

Modified the J2MiniSolver to user the new interface.

Modified the ParallelNeohookean solver to use the new interface.

Added a convenience make_unique function that is identical to the C++14 make_unique