start renaming MeshSizeField to MeshAdaptMethod
since this class has evolved to the point where
it calls ma::adapt(), and will continue to grow
to the point where it transforms meshes significantly
before and after adapting, it can no longer in good
conscience be called a "size field".

renamed the class and its pure virtual methods to
reflect their ever-expanding role.