PUMI: overhaul the field size logic
there are many separate locations where
attempts are being made to determine
the number of entries per node that are
stored in a Vector.
This commit is the first step towards
consolidating this logic, because our
FELIX case requires more flexibility
(i.e. 2D vector fields on a 3D mesh)
and the copy-paste state of the code
would have required much redudant extension.