Some changes in the FELIX problems, plus new evaluator
We update the problems according to the changes in the evaluators.
In particular, the problems need to create SharedParameter evaluators
to make the Homotopy parameter available. Moreover, some evaluators
are now templated on booleans, specifying what kind of problem we
have, so the creation of those evaluators had to be updated.

We also added an evaluator for Dirichlet BC for the Hydrology problem.