Change in Aeras_SpectralDiscretization.hpp
We again removed the useless attributes related to side set
structures. Instead, we return a reference to a statically
created object inside the getter. Notice that these functions
should never return and throw an exception instead.

Not having the useless attributes keeps the mesh structure
cleaner, and easier to maintain.