CUDA Bugfix: Added device view for PHAL_GatherSolution
An illegal memory access is encountered in PHAL_GatherSolution,
PHAL_ScatterResidual and Aeras_ShallowWaterResid when running with
Kokkos::Cuda on GPUs for Aeras ShallowWater problems. This was fixed in
PHAL_GatherSolution by using Tpetra, getLocalView() to obtain the local
data on a specific device. PHAL_ScatterResidual and
Aeras_ShallowWaterResid still needs to be fixed so a preprocessor
directive was used to run the serial code when using Kokkos::Cuda.