explicitly instantiate ParallelConstitutiveModel
First off, this silences compiler warnings of the
form:

J2MiniSolver.hpp:85:8: warning: inline function
‘void LCM::J2MiniKernel::operator()(int, int) const
[with EvalT = PHAL::AlbanyTraits::DistParamDeriv; Traits = PHAL::AlbanyTraits]’
used but never defined [enabled by default]

And will hopefully help with #9

The reason for the above warning was that
ParallelConstitutiveModel<...>::computeState(...)
was being implicitly instantiated, and it
does a Kokkos::parallel_for using
Kernel::operator()(...).
However, for all three values of Kernel,
operator()(...) was only defined in the
explicit instantiation in the .cpp file for said Kernel.
For normal host execution, this linkage can be
resolved, but ideally code that runs in a
Kokkos::parallel_for should be visible in the same translation unit.
More importantly, implicit instantiation of
ParallelConstitutiveModel seems wrong, so the
first thing to do was prevent that from happening.
It turns out ParallelConstitutiveModel.hpp includes
ParallelConstitutiveModel_Def.hpp which is backwards
and defeats the purpose of a _Def.hpp file.
Removing that include prevents implicit instantiation.

Next we need to explicitly instantiate the uses of
ParallelConstitutiveModel in the .cpp files corresponding
to where they are defined as template aliases.
However, doing so was a bit difficult.
PHAL_INSTANTIATE_TEMPLATE_CLASS can't handle the
extra Kernel argument.
PHAL_INSTANTIATE_TEMPLATE_CLASS_WITH_EXTRA_ARGS
Also can't do it directly because Kernel is itself
templated on EvalT and Traits.
PHAL_INSTANTIATE_TEMPLATE_CLASS can't be used
on the template alias (e.g. J2MiniSolver) because
template aliases cannot be explicitly instantiated:
http://en.cppreference.com/w/cpp/language/type_alias
http://stackoverflow.com/questions/25118191/explicitly-instantiate-class-through-template-alias
For these reasons we replace the template aliases with
simple template classes that derive from specializations
of ParallelConstitutiveModel.
These classes can be passed to PHAL_INSTANTIATE_TEMPLATE_CLASS.
Their constructors must be defined in the .cpp file as well
to prevent undefined references to them elsewhere.