Introduced a FELIX problem factory
The albany problem factory queries the FELIX one for the
creation of the problem (if needed), and the FELIX factory
takes care of creating the problem. This limits the pollution
of FELIX code into the main albany code. For instance, avoid
recompilation of Albany_ProbleFactory if a new FELIX problem
is added, which would be pointless for people not even enabling
FELIX.