Small changes in some evaluators in FELIX
The evaluators that can be used in StokesFO, Hydrology and
StokesFOHydrology have one/two extra boolean template arguments
specifying the type of problem. Moreover, for the Hydrology, we
also distinguished between the case where we solved only the
Hydro-Potential equation or also the Water Thickness equation.
Before, we provided a partial specialization for all cases. Now
we avoid that, using std::conditional where possible to select
the proper Scalar type, or we simply use a switch in the method
evaluateFields, to distinguish between sideset-based and cell-based
evaluations.