Aeras CUDA: Removed large memory allocation in ComputeAndScatterJac
Large temporary data structures were created in the buildLaplace kernel
in ComputeAndScatterJac to avoid data corruption on the GPU. This has
been changed to static arrays which are created on the fly. A switch is
created for numNodes so that the size of the array is known at compile
time.