LCM: Remove LCMUtils files. Remove FloatingPoint files.
The utilities files were very small and the functions there
are better placed in MaterialDatabase. Others were not used at all.

The portable FloatingPoing class to enable and disable trapping
of floating point exceptions is better suited for MiniTensor, so
it will be moved there.

Other changes to reflect this.