LCM: Add ability to use ROL's optimization algorithms with MiniSolver.
- Add optional Albany dependency on ROL. If Trilinos is compiled
with ROL, ROL-enabled MiniSolver examples and tests will be enabled too.
- Add ROL-enabled minimizer class that mimics the "native" MiniSolver
minimizer class as much as possible. Among the differences is that
ROL requires options to algorithms be passed as Teuchos::ParameterList.
- Add ability to handle Albany type traits for ROL-enabled MiniSolver.
Just as the "native" MiniSolver implementation uses Sacado to
compute gradients and Hessians, the ROL implementation does this too.
This requires handling of the Albany type traits in EvalT so that
the FAD types from Albany and MiniSolver are not mixed and confused.
- Add first test for all this.