LCM: Fix original Schwarz BC after refactor for Kokkos DynamicRankView.
After the Intrepid2 refactor for Kokkos DynamicRankView, the API to
some of the functions in CellTools changed and require Kokkos views
of different ranks for the same objects. This produced an error in
the manipulation of physical and parametric coordinates for the
interpolation of the original, serial Schwarz boundary condition.
Ranks were changed in some instances and temporary objects created
in others to fix this. This Schwarz BC is preserved for debugging
and verification purposes against the more general DTK-based one.