separate updateMesh calls for each discretization
The addition of several solution transfer features
to the APF discretization, most importantly updating
LCM Time values (commit faf075d), resulted in
several overloads of updateMesh(), and created
a dangerous situation whereby developers could call
the wrong one and important data would not be transferred,
as happened in commit e370bbb8.
This commit removes the virtual updateMesh()
declaration, because:
1) it is too tempting (and wrong) to call when using APF
2) in all current Albany code it is clear whether STK or APF
is being used, and the derived C++ type is already used
3) this allows STKDiscretization::updateMesh to be
called within the constructor, because it is no longer
virtual (let me know if I messed up the STK changes).