Creating Rythmos and Tempus RK4 tests for TransientHeat2D.  Rythmos and Tempus give the same solutions, which is more evidence Piro::TempusSolver is working properly. Also, scratch the comment in my last commit about the last time step in the Exodus file being incorrect with Tempus.  It is actually correct.
Note that it appears to get the correct solution, an implicit scheme must be used
for this example (an explicit scheme gives a different solution than an implicit one).
Hence, we may want to do further testing of the TempusSolver where an explicit
time integration scheme gives the right solution.