Adding another test for Tempus time-integration scheme with explicit RK4 time-stepping. Results match Rythmos results.
It appears TempusSolver works with implicit and explicit schemes, with the exception
of problems on spherical grids (e.g., Aeras shallow water and 3D hydrostatic problems),
where the tempus solution blows up.  This is not critical given that Tempus will be
primarily used for LCM problems; but should talk to Curt Ober about what seems to be
the issue for the spherical grids.