tracking down an issue with IPtoNodalField
fundamentally, the issue is that it has
this manager tracking its instances, but
it really only wants to do that for Residual
specializations, and essentially ignore
all other specializations.
this mismatch in expectations is what causes
nodal_Cauchy_Stress to not show up in the
output file when expected.
this commit tries to move the manager logic
into only the Residual specialization,
however it results in a segfault.
more work to be done...