Albany_RBGen:  fix issue causing segfaults
The issue has to do with whether ScalarValueState is a vector of strings
or a vector of pointers to strings.

In April 2015, ScalarValueState was changed to a vector or pointers to
strings.  The commit message indicates that this was done for issues
relating to Cuda.

The current commit changes ScalarValueState to a vector of strings if
Albany is built with -D ENABLE_MOR=ON, but leaves it as a vector of
pointers if -D ENABLE_MOR=OFF.

This means that Albany built with -D ENABLE_MOR=ON may not work with Cuda.

As discussed above, the current commit undoes some of the changes
made in:

    commit 8300e056c870651871e87d4d3cd710539526bbf8
    Author: Glen Hansen <gahanse@sandia.gov>
    Date:   Fri Apr 3 09:23:07 2015 -0600

        Missed an edit needed to dereference the std::string *

        Missed an edit needed to dereference the std::string *. Changed
        the storage of the time variable in the state manager to use
        a string * instead of a string to make nvcc happy.

    commit 5094b0ac35e1ac0c06abfaaf4f6e60c0b6968c0e
    Author: Glen Hansen <gahanse@sandia.gov>
    Date:   Fri Apr 3 04:50:58 2015 +0000

        Fix some Shannon CUDA compile issues.

        Add some ifdef's to turn off Teko when it is
        not enabled in Trilinos.