This commit fixes an issue caused by a previous commit of mine.
In commit 45ba230091a45af532998037682ec908607b93da
ScalarValueState was changed to be a vector of strings instead of a
vector of pointers to strings.  This change is only made when Albany
is built with -D ENABLE_MOR=ON.

This commit applies this logic to an additional source file that I
missed in the original commit.

Again, this change to ScalarValueState may mean that Albany can not
be built with Cuda if -D ENABLE_MOR=ON