LCM: adding parameterlist option to turn on/off registering of dirichlet_field for Mechanics problems.  It is on by default, as was the case before, but can be turned off by adding the line
<Parameter name="Register dirichlet_field" type="bool" value="false" />

to an input file; please see examples/LCM/MechWithHydrogenFastPath/surface_diffusion/surfaceDiffusion.xml

for an example of usage.

Turning off registration of dirichlet_field for MechWithHydrogenFastPath/surface_diffusion
example.  Test still fails on my machine, so I believe it is something other than the
presense of the dirichlet_field that is causing the test to fail.