Removing Epetra overlapped_dgdp member variable from PHAL::Workset and replacing it with Tpetra analog (overlapped_dgdpT).  This allowed conversion of ATO responses to Tpetra.
One caveat is there are ifdefs throughout the code to build distributed
responses only if ALBANY_EPETRA_EXE is on.  Hence, if one tries to run
ATO tests with distributed responses with Epetra turned off, bad things
may happen.  If distributed responses are meant
to work with AlbanyT executable, the ALBANY_EPETRA_EXE ifdefs pertaining to
them throughout Albany should be removed.  I will submit this as an issue shortly.

Another caveat is I cannot seem to switch workset.dgdp to workset.dgdpT in the code -
when I try to make this trivial change, the ATO::Constraint_2D_adj test fails.
This is not key right now, but is needed to remove dependency on Epetra of the distributed
responses further down the line.  I will submit this as an Albany issue on github
too.